Settlers and settling cities
Created the basic system for settlers and creating cities. The UI for it isn't in place (as Joel is still working on the Unit UI), but as long as you have a settler selected you are now able to settle a city in any tile that they're able to enter.